revent: Add addListenerByName(), improve event priority sorting
You can now register for a typed event by name rather than by its actual
type, which may save you from having to import a module just to get the
type.
Also, the system for ordering handlers by priority did not used to be able
to force a low priority when other handlers were using the default
"dontcare" priority.  It can now.  This slightly breaks the old semantics,
but only a corner case that was sort of ill-defined anyway.  Your choices
now are:
 * All handlers are dontcare (None) and are fired in order added
 * All handlers run in priorty order, and None is the same as 0