libopenflow: Clean up, remove redundancy, etc.
The autogenerated constants, lists of constant names, and dicts of constants
were redundant and sometimes nonsense (some of the dicts had duplicate keys,
some of the lists were useless).  There is now a single "reverse dictionary"
for all of these, which is used to generate global variables, lists (when
appropriate), and "forward dictionaries".