recoco: Improve behavior when cancelling
If the callback cancels a recurring timer, it now exits immediately instead
of cycling another time and then cancelling before calling the callback.
Also, if you cancel it before it actually ran, it will now quit immediately.
Really, though, it'd be nice to be able to interrupt a waiting timer when
it gets cancelled.