log_messenger: Keep the listener IDs.
This was done because the GUI messenger service needs to rip them off.  There
are other ways this might have been accomplished (like not adding them in the
first place), and it might be better if messenger just had more facilities
for multiplexing connections, but this will work for now.