Revert previous commit "'message' value should be record.msg (?)"
The original behavior was intentional.  A formatted version of the log
record is probably the single most useful thing to do with logging, so we
definitely want to send that.  When json=False, this is all we send.  If we
just send record.msg (as the previous commit does), this means that a
client would never be able to see the formatted version;  a specified
"format" string would never do anything at all, and any user arguments to
the log function (e.g., 'log.debug("foo %s", bar)' would never be
substituted (and couldn't possibly be subsituted by the client because
when json=False, record.args are never sent to the client in any form).
Even when json=True, the exact behavior of formatting on the POX side can
not be replicated on a client side as the .args have already been converted
to strings.
So we ALWAYS want to send a formatted version.  Indeed, I could not think
of a reason why the unformatted version (record.msg) would ever be
valuable, which is why I didn't included it.  I called the formatted
version "message" rather than "msg" to avoid confusion between the two.
If the unformatted version is ever useful for some reason, we can send it
too by just add 'msg' to _attributes.