log_messenger: Better handling of record.args
Previously, all .args were just forced to strings.  Now they should retain
their types if they are JSONable types and only be forced to strings if
they cannot otherwise be JSONified.  Note this means that circular
references in arguments passed to log functions will make this choke.  If
that is ever a problem in practice, let me know.
This patch would make the record.msg field considerably more useful,
though I am still not convinced there is ever an actual use for it. :)