packet: Fix IP identification field
The default value for this field was int(time.time()), which is not really a
very unique value, and also is too large to actually fit the two bytes it
gets packed into.

It looks like at one time, this was just filled in from a counter and the
code was never entirely removed.  I have ressurected this, but initialize it
to the current time and make sure the value fits in two bytes.  Hopefully
there wasn't a really good reason to not use it that I am missing. :)