libopenflow: flow_mod buffer_id defaults to -1
Previously, buffer_id defaulted to 0, which means you had to set it
explicitly if you didn't want to use it!

Additionally, 0xffFFffFF is now mapped back and forth to -1 for
(hopefully) all buffer_ids.  So you should use -1 and expect -1 (which
is what it says in the spec).

Lastly, when doing an output action to the controller, it defaults to
sending the entire packet.