packet: hdr() knows payload length; DHCP options are updated
Hopefully hdr() is now called with the length of its payload as an
argument, letting it update its length field, if any.  Hopefully
this now works for at least ipv4, and udp.

Additionally, if you alter the parsed options in DHCP, it will now
hopefully reflect those changes when pack()ed.