Add multiple-instance feature for launch()
There are cases where you may want to specify a module more than once on the
commandline.  This used to always cause an assertion failure.  Now you can
write launch() functions that handle this.  If you include a parameter called
__INSTANCE__, then you'll be able to specify the module multiple times.
When launch() is called, it will be passed a tuple for __INSTANCE__:
(instanceNum, numInstances, last), where instanceNum is [0..numInstances), and
last is True the last time it's called (when instanceNum = numInstances - 1).