openflow: Use NONE in_port instead of CONTROLLER
The default port for packet_out messages used to be OFPP_CONTROLLER, which
sort of makes sense for synthesized packets. This used to work fine, but
recent versions of Open vSwitch stopped supporting it.  The only logical
alternative if OFPP_NONE, which ALSO stopped working in Open vSwitch until
Sept 8 / 2011.  Since this works, it's now the default.

Discovery used to have OFPP_CONTROLLER hardcoded for the LLDP messages it
sends.  It now just uses whatever the default is.