Changes to pox.py for PyPy
If you put a PyPy 1.6 or above distribution next to pox.py using "pypy"
as its top-level name, then pox.py will run using it.

Additionally, readline in PyPy has a nasty effect on the log messages that
POX spits out because it turns off terminal output postprocessing (so
newline -> CRLF translation is disabled).  We monkeypatch to turn it back
on which doesn't immediately seem to have any terrible results and fixes
the annoying problem with the log messages.