packet: Many changes
This is a fairly substantial revision to the packet library and will without
question introduce bugs and incompatibilities.  Some of the goals:

* Make packet construction much easier:
  * Most packets take keyword arguments in their constructors for most fields
  * Build more packets from parsed fields rather than raw unparsed data
* Improve DNS classes
* Use EthAddr/IPAddr more places (e.g., DNS)
* Remove more old array legacy:
  * Fix bugs from things still expecting an array instead of a bytes
  * Change "arr" to "raw"
* Add unpack() class method, which is probably to become the preferred way
  to initialize packet objects from data
* Fix a number of bugs

Sorry this isn't better documented or tested, but I've had it in a separate
branch for weeks and it hasn't been moving, so I just wanted to get it pushed.
Please report bugs, and don't hesistate to roll back your own branch for the
time being.