packet: packet classes now call base class __init__()
Instances of packet_base subclasses used to get a handful of instance
variables from the packet_base *class* variables.  I sort of discourage
this in general, especially when they have no meaning on the base class
and are only used as a shortcut for getting instance variables for free.
In addition, this can be blamed for a huge recently-fixed bug.

So, all instance variables on packet_base are now created explicitly in
packet_base.__init__(), and all subclasses explicitly call this.