messenger: Fix typo in log service
I fixed this same typo in messenger itself a while back, but never
updated the log service because it was all going to get fixed when
I pushed the updated messenger stuff anyway.  Since I haven't found
the time to finish and push the updated messenger stuff for weeks,
I might as well make it all at least work again.