revent: Adjusted handler prioritizing; improved documentation
Handler prioritization had a couple revisions before and was subtle.
It had some gotchas (done with performance in mind) which were not
commented.  It should now be more straightforward.  Specifically, it
should be the case that:
* For a list of handlers which are all unspecified priorities,
  the list is never sorted.
* Whenever you add a handler with a priority, the list is re-sorted.
  In part this is true because...
* An unspecified (None) priority is less than any specified priority.
  (Kept true by sorting specified ones and appending unspecified to end)

Additionally, I've tweaked/fixed/filled in some of Colin's documentation
and comments.

Also other minor tweaks and fixes.