core: Revert changeset 408
This change was Colin posing the question of why callLater()
and raiseLater() used "_self" instead of "self".  The answer is
that one might want to specify something named "self" as a
keyword argument and I figure the chances of wanting to specify
something named "_self" as a keyword argument are lower.

Question answered; comment removed.