pox.lib.pylibopenflow: Adding a new module to pox.lib
For testing, I need to simulate the behavior of a switch. Stanford's
pylibopenflow contains a python (messaging) simulator for switches. I ported
this to pox, and added it to pox.lib. Feel free to move it to a different
location if need be.

In reality, openflow messages may be the wrong layer of abstraction for
simulating the behavior of a switch. In future versions of the NOM, we don't
plan to expose any openflow specifics to the application. Rather, we plan to
expose some sort of "Forwarding Engine" abstraction that encapsualtes the
routing tables of the switch, but doesn't expose openflow messages themsleves.

Nonetheless, pylibopenflow might help serve as a blueprint for implmenting the
"Forwarding Engine" abstraction.