of_01: revert r457
of_01 needs openflow to run -- there's no point checking in
individual places to see if openflow is loaded.  A better way
to resolve this is to have of_01 listen for the openflow
module to come up (perhaps using the component base class).

Also, the assert concerning a connection's buffer should not
be there -- the point to the buffer is that an OpenFlow
message may span multiple reads.