recoco: Making scheduling safer
I think Scheduler.schedule() and the default behavior for
Task.start() should now protect you from scheduling a Task
more than once, even from multiple threads.

There is a cost associated with this, so there is now a fast
scheduling option as well (Task.start(fast=True) and
Scheduler.fast_schedule()).

Task also gets a new factory method which automatically starts
new tasks with fast=True.

Lastly, the sanity check for scheduling (which is now in
fast_schedule()) is an assert, so it will be removed when
running in optimized mode.