openflow: Add spanning_tree module
This is a rough first pass at an OpenFlow spanning tree
module.  It is conceptually similar to the NOX spanning
tree component.

Right now it responds to openflow.discovery link messages.
When there is a change, it builds a spanning tree and then
turns off flooding for all switch-to-switch ports that are
not along the tree.

Not really tested, so, you know... enjoy.