libopenflow: Revert r478
This patch added print statements in unpack() when passed not
enough data.  I don't think this condition necessarily represents
a problem, though.  If it does, I think the caller can notice it
and print their own warning by looking at the return value.

If a change is warranted, I think it's maybe to make this
condition raise an exception.  Would be happy to discuss.