l2_multi: Floods more
Before, if a packet came in on a switch port and there was no flow,
it would just be dropped.  This is too conservative for when the
packet has been flooded because we don't know the destination --
it would only be flooded out the first switch.  So now we are
willing to install flows for these "lost" packets.

One more improvement it'd be nice to make is to check the spanning
tree (if it's running) to see if the port the lost packet arrived
on is in the spanning tree.  If it isn't, we should be able to
drop it.

--HG--
extra : rebase_source : 52408d478473983e745553bf975778f95c3b6953