topology: Kyriakos pointed out another id=None missing arg bug
Should work now.

The motivation for these changes is that because Entity.id has to be unique,
using OpenFlowSwitch's dpid as a id is somewhat dangerous. Suppose another
entity (say, Host) is added to topology first. Later, an OpenFlowSwitch tries
to add itself using its' dpid, which has the same value as the Host's id. At
this point we're somewhat screwed -- we cant change Host's id, since the user
may have a reference to it. I suppose we could fall back to a normal id for
the switch, but that seems inconsistent.