topology.topology: Reserve IDs 1-100 for mininet switches
This is just a temp patch to avoid ID overlapping between mininet dpids and IDs
assinged to ports. This is only to keep things working until we agree on an ID
scheme. So for now at least mininet topologies with max 100 switches wil work,
maintaining the property id==dpid for switches.