of_01: Fix for platforms without select.PIPE_BUF
This value is used for slicing up output chunks, but it's not
available on all platforms.  This patch looks for it where
PyPy sometimes hides it, and otherwise just uses a hardcoded
default value (which is compatible with POSIX).

The DeferedSender mechanism (of which this is part) is seldom
actually invoked (explaining why nobody noticed this before),
and will probably be removed once the ioworker stuff gets a
bit more mature.

This patch is based on one by apanda.