openflow: Add workaround for HP switches
HP switchies don't support barrier messages (which I am pretty
sure are not optional...), but POX's connection procedure relied
on them to decide when a connection was "up".  The result was
that POX wouldn't accept connections from HP switches.

This commit is a sort of ridiculous workaround.  POX still sends
a barrier request, but instead of just looking for the
corresponding barrier reply, it also will accept a corresponding
OpenFlow error message (i.e., with the same xid).  The error
will still be logged, but POX will accept the connection.