l2_learning: Don't try to drop nonexistent buffers
(Resolves issue #9)

l2_learning's drop() function is sometimes called simply to
make the switch consume a buffer by sending a packet_out. BUT,
this is called regardless of whether the switch buffered the
packet or not (i.e., if buffer_id is -1 or not). If there's no
buffer, there's no point in doing this to begin with AND the
switch will probably return an error, so in this case we just
shouldn't do it. :)