l2_learning: Don't try to flood unbuffered packets
This is similar to that raised in Issue #9.  If we want to
flood a packet, but it isn't buffered at the switch, we
no longer try.  We COULD try to flood it based on the data
we got from the packet_in, but it might be incomplete so
we just don't.