openflow: Fix error when switches disconnect unexpectedly
This closes Issue #20 and Issue #18.

The disconnect code logic was backwards and also a bit broken
(probably as a result of trying to "fix" the fact that it was
backwards).  The notable effect of this was that occasionally
POX would stop connecting to switches when one unexpectedly
went down.

This commit tries to make the disconnect code clearer and
actually correct.