libopenflow: Fix two minor stats-related bugs
This closes Issue #16.

One of these is a regression from 01f83795a, which changed the name
of ofp_aggregate_stats_reply to ofp_aggregate_stats.  The latter is
how it appears in the spec, but we explicitly decided to disregard
that.  Both names now work.  (But use the _reply one.)

The other seems to just have been a bug.