openflow: Un-refactor ConnectionArbiter/Nexus
fdbd113d took ConnectionArbiter, OpenFlowNexus, and launch out
of openflow/__init__ and put them in connection_arbiter.py so
that __init__ could be loaded by itself (for the SDN debugger).
This has the effect of making connection_arbiter the real
OpenFlow "component", though, which isn't ideal.  Since I
think the only thing that was problematic was that it depended
on pox.core, I have just tweaked that aspect of it and stuck
it back in openflow/__init__.