lib.graph: Add in dummy Node class
The Node class was removed as part of some cleanup Panda did, though
it's still referenced in one place.  Additionally, it may still be
nice to have an OPTIONAL node class with convenience methods that
just call methods on the containing graph (modeled after the
original methods back in like b4502d422).

This commit just adds back a dummy Node class with a note, which at
least has the effect of making all tests pass.