graph.util: Add JSON encoder/decoder
Used for (de)serialization of custom python objects. The encoder
traverses the objects' dictionaries creates a json string that
describes the objects, classes, and modules inside the serialized
object. The creates new objects as necessary, importing required
modules etc when needed. The goal is to work for custom classes that
potentially have custom objects as members.