openflow: Add ability to suppress OpenFlow errors
Previously, whenever an ofp_error_msg was received, it would be
logged.  Now, a listener can set the event's should_log to False
to suppress this.

It's possible that this should be implemented differently -- by
just having the error logging part be a normal listener at a low
priority (as it is, it's a special case).