core: Fix issue with dependency startup
Because dependencies can be resolved sort of recursively, we
sometimes had a case where we try to resolve a dependency
multiple times and it works multiple times so we call the
notification callback more than once.  We now check to
make sure it hasn't been satisfied before we try agian.
(A better solution might be to not do it recursively.)