Remove distribution stuff
So this hasn't been synced with its upstream in a while, nor is
its upstream tracking betta as far as I know.  The files need to
get moved around anyway, and it's possible these should be in an
"ext" project anyway.

SO, I am removing them for now.  We can potentially bring them
back later, but it might be easier for them to be maintained
elsewhere anyway.