libopenflow: Rework resend mechanism
I think this more elegant and unified.  You can forward the data
from a packet_in by specifying the packet_in as the data for a
packet_out or a flow_mod.  In the case of a flow_mod, this may
result in pack()ing the flow_mod actually being a flow_mod,
a barrier, and a packet_out.