libopenflow: Implement len() for all structs
You can now use len() on any ofp_base subclass and any
instance of any ofp_base subclass.  The former returns
the minimum length, the latter should return the
packed length.  For fixed-size structures, these are
the same.

This is a step on the road to removing the length
attribute.