tk: Make signals work
When the main thread is stuck in the Tk loop, signals aren't
handled, since they're handled by the Python interpreter.
This commit causes the Tk loop to jump back out to Python
twice a second so that signals will be handled.

In particular, this makes Ctrl-C from the commandline kill
POX reliably again even when using Tk.