openflow: Make openflow.connections special dict
This makes openflow.connections not just a list, but a special
dict.  For example, if you iterate over it, you get the values
instead of the keys by default, which is more useful.  "in"
tests for both DPIDs and actual connections, and there's a
nicely named method for iterating DPIDs.