openflow: Major changes to libopenflow, etc.
This is a pretty substantial patch for libopenflow and some of the
periphery stuff (e.g., of_01).  While things could always be better,
this hopefully more or less finishes the mission I started at the
end of September to clean up a lot of things that had been bothering
me about libopenflow.  The major focus of this commit is unpacking,
but here's a somewhat more complete list:

* It's finally line-wrapped nicely.
* I think there are now packers and unpackers for everything (there
  had been a few missing here and there before).
* Better and more consistent error handling for unpack problems.
* Better interface for unpacking.
* You can unpack out of a buffer at an arbitrary offset.
* A number of smallish assorted bug fixes.
* Some badly named items have been tweaked.
* Support for queue properties.
* All vestiges of storing the length from the message header are now
  gone -- all lengths are computed.