packet: Reintroduce 802.2 LLC/SNAP
I don't think 802.2 worked in any released version of POX (and
maybe not even any released NOX), and the llc file was removed
long ago.  This is a total rewrite and should more or less work.

Additionally, VLAN-in-VLAN is now supported, and the
effective_ethertype should now be more consistent, even though
it should really be called inner_ethertype... and for that
matter, when it's LLC, it's not even necessarily an ethertype,
but whatever.