datapath: Fix some output stuff
Output to a bad port number isn't an exception -- it's just a warning.
This is because it's the controller which has the problem; we just have
to deal with it -- which we can (by ignoring it).

Also, we were previously logging a message when outputting to a down port,
but outputting anyway.  This wasn't a problem, since we currently pretty
much always have LINK_DOWN too when a port is down, but it was wrong and
is now fixed.