recoco: Fixes to scheduling priorities
Peter Peresini pointed out that the check for t >= 1 was
nonsense.  It was *supposed* to be an early-out for
priority==1, which it now is.

In the process of fixing that, I noticed that it looked
like the priority check was backwards and like a slightly
different check on the ready list would also make sense.