of_01: Alter disconnect event
When a disconnection occurs during an operation (e.g., send()),
we now defer the disconnect event so that it happens later
(sometime not during handling of an event).  This prevents
some ugly stuff from happening (e.g., connections disappearing).

It's possible there should be separate "disconnect" and "reap"
events or something...