of_01: Add eth_addr property to Connection object
This uses the DPID to figure an Ethernet address, as suggested by the
OpenFlow spec.  An alternate approach might be to use the local port's
address, but I don't think we're guaranteed to have one.

There are multiple places where we could refactor code to use this
instead of a helper function.

This also fixes a latent bug since EthAddr wasn't imported.