nx: Add OFPC_INVALID_TTL_TO_CONTROLLER const
This is actually from OF 1.1 or something, but it can be used with OVS.
However, it's deprecated in OF 1.3, I think and isn't strictly necessary
since you can get around it with a Nicira extension and probably some new
OF feature too.  Bottom line, including it is questionable, but I'm doing
it anyway (for now).