ioworkers: Fix Persistent worker sync connection
If a persistent IO worker's connection failed immediately (which
I guess I never saw happen on Linux or Mac OS), it raises an
exception from the constructor -- a difference in the API from
when it fails asynchronously.  It should now act the same
either way.

Closes #67.  (Hopefully.)

Thanks to damomeen for the report.