log: Fix --no-default option
This is a regression caused by d8b65a739a5f, which turns hyphens
into underscores to make them compatible with actual argument
names in the launch function.  In this particular case, it's
not necessary, because no-default is put into kwargs and CAN
have hyphens, but boot doesn't take that into account.

It could, though.  And maybe it should.  At this point, boot
could use some work in general (e.g., refactoring and cleanup
based on the new help component), so maybe we should come back
to this at some point.

Thanks to Alison Chan for the bug report.