nx: Add NXM size-hinting
Formerly, things like register moves could guess the number of bits to
copy, but did so based on the number of bytes in the nxm_entry (just
multiplying it by 8).  This doesn't work for fields with sizes that
aren't multiples of 8, so they had to be specified by hand.  This lets an
nxm_entry subclass set _size_hint or override _get_size_hint() to
make it whatever it wants.