nx: Make nx_match easier to use
It can be confusing when you set an attribute on nx_match which doesn't
actually correspond to an NXM type.  You end up just creating a worthless
attribute which doesn't actually contribute to the match and which doesn't
generate any error.

This patch "locks" the nx_match after it's created, so trying to write new
attributes will raise an exception, alerting one to the fact that one's
intention is not going to be honored.

Note: We have seen similar bugs in the wild with ofp_match too -- it might
make sense to do something similar there as well.