nx: Don't alter attribute name in nx_match
Previously, we used an attempted "fixed" attribute name in the fall-through
case of nx_match (when there's no corresponding match type), so either
a new attribute may end up with a weird name, or the exception text
referred to a name that wasn't the one the user actually used.  Fixed.