datapaths: Refer to capabilities by right names
The switch was initializing its Features object using attribute names
which aren't the actual significant attribute names (it was missing the
'cap_' prefix).  We now both fix this and make it harder to do it without
noticing by having the Features object override __setattr__ and complain
when you're doing it.