datapaths: Add port counters
Unfortunately, this is currently sort of expensive.  At the appropriate
places, all we have are packet objects.  We need to pack these so that
we can find the lengths, and then we probably just throw away the packed
result even though it may be used again soon (and somewhere, we may have
already had the packed version).  So the efficiency here could be
improved.

rx_packet() now takes an optional argument for the packed version.  We
don't currently use it, but this would allow us to skip an extra pack.
In the future, we should continue this sort of thing, I guess (carrying
the packed packet too when possible).

Thanks to eternia478 and nbjsw for the bug report.