packet: Adjust conditions for not-packing unparsed packets
1705997e3 added logic to packet_base which stops unparsed packets from
attempting to re-pack, and just uses their .raw instead.  It assumed that
.next would be set to None in cases where this should happen, but this
isn't the case with ipv4, for example, which has it set to an empty bytes.

This patch now counts anything falsey for .next as okay for non-repacking.
It's possible that this isn't a great idea and that things like ipv4 should
explicitly set .next to None instead.