packet: Set .next to None when parsing fails
This is related to 029b041 and 1705997.  We now try to make sure that
when parsing fails, .next is set to None.  This is already the case
for most packet classes, as they initialize .next to None.  Some
classes, however, initialize to an empty bytes object.  We now set
.next to None at the start of parsing for these classes.

It's possible that .next should always just be initialized to None,
but I couldn't convince myself this wouldn't break anything, so I'm
leaving that for another day.