datapaths: Fix buffer_id in log message
The buffer_ids we send to the controller are 1-based instead of 0-based
(as far as I can tell this is for no particular reason, but I don't mind
it, especially since it turns up bugs in OFTest).  We actually store the
buffers in a 0-based list, so we increment/decrement the buffer_id as
appropriate.

However, when logging problems involving the buffer_id, we were logging
the problem using the decremented value -- not the one actually passed
over the wire.  This was confusing (especially in the case of 0 since it
now looks like -1) and has been fixed.