datapaths: Implemented queue stats requests
Since the switch doesn't actually have any queues, upon requesting
statistics on all queues, an empty list will be returned to be packed
and sent back.

As of yet, no checks against port_no are currently implemented, no
matter if port_no is invalid or not. On the other hand, if the queue_id
is invalid, then an approximate error would be returned.