datapaths: Fix issues from moving process_flow_mod
Some issues came up after the last commit:
1. The table can't be defaulted to self.table. It may be possible to
just check afterwards if it is None or not (and set it), but this seems
a bit too hacky. As such, both connection and table have default values
removed. It should be better this ways, since both cannot afford to be
any sort of default values.
2. All references to self are changed to table, as previously discussed
about changes. table._table is also changed to table.entries
3. To not require importing more than necessary, adding of an entry via
a flow_mod is now implemented in flow_table.py itself, in the FlowTable
class as method add_flow_mod_entry. A question arises on whether or not
to refactor this into add_entry itself, via a mere check that it already
does.