datapaths: Implement flow deletion port filtering
Added code in the flow deletion option to handle parameter out_port.
Removed the NotImplemented Error associated with the fact that checking
against the out_port when deleting an entry is not implemented.

This required adding a field to FlowTable's remove_matching_entries
in order to allow filtering by out_port.