datapaths: Fix check overlap NotImplementedError
Implemented the check for overlap if the OFPFF_CHECK_OVERLAP flag is
set to be on. Before any attempt to add, extraction of matching packets
and comparison with priorities is done to see if there is any overlap
with the current to-be-added entry.

From both test detailed_contr_sw_messages.OverlapChecking and from the
specs, it seems that the matching should be allowed to go either ways.
Unfortunately, this means it is impossible to just use matching_entries
to find overlaps, and is_matched_by would require both sides to be of
the TableEntry class. Hence, we can at best just directly use the
matches_with_wildcards method to do match, while matching on priority
as well while traversing the table.

There are 2 things of concern now:
1. Because we are forced to manually traverse by priority, it makes any
optimization based on priority easier. Since the table is sorted, we
should just look at whichever is actually of the same priority.
2. _process_flow_mod at this point has become very cluttered. I was
thinking of refactoring some code out, i.e. the adding part and the
overlay checking part. Also, if we can just organize it a bit more...