datapaths: Implement table entry overlap checking
Implemented the check for overlap if the OFPFF_CHECK_OVERLAP flag is
set to be on. Before the attempt to add an entry, matching and priority
comparison is done to see if there is any overlap with the current
to-be-added entry. Removed the NotImplemented Error about this.

From both test detailed_contr_sw_messages.OverlapChecking and from the
specs, it seems that the matching should be allowed to go either ways.
Unfortunately, this means it is impossible to just use matching_entries
to find overlaps, and is_matched_by would require both sides to be of
the TableEntry class. Along with requiring comparison of effective
priorities, the flow mod is already converted to a table entry to ease
the comparison. This results in reverting the change that led to the
creation of add_flow_mod_entry 4 commits back (fix from moving the
process_flow_mod method to switch.py).

There are 3 things of concern here:
1. There is an ambiguity in the specs of whether priority matching is
based on effective priority or the priority of the flow mod itself. Due
to the fact that the spec does not clarify any such difference between
the two and the fact that packet matching only considers effective
priorities, it was chosen that the overlap checking would also be only
on the effective priority.
2. We assume that the flow table entries are sorted in decreasing order
by effective priorities. If this changes, the fact that the overlap
checking skips over irrelevant priorities would cause problems.
3. _process_flow_mod at this point has become very cluttered. I was
thinking of refactoring some code out, i.e. the adding part and the
overlay checking part. Also, if we can just organize it a bit more...