datapaths: Implement table entry overlap checking
Implemented the check for overlap if the OFPFF_CHECK_OVERLAP flag is
set to be on. Before the attempt to add an entry, matching and priority
comparison is done to see if there is any overlap with the current
to-be-added entry. Removed the NotImplemented Error about this.

From both test detailed_contr_sw_messages.OverlapChecking and from the
specs, it seems that the matching should be allowed to go either ways.
Unfortunately, this means it is impossible to just use matching_entries
to find overlaps, and is_matched_by would require both sides to be of
the TableEntry class. Along with requiring comparison of effective
priorities, the flow mod is already converted to a table entry to ease
the comparison.