datapaths: Fix non-spec command AssertionError
For the final case of processing a flow_mod, in case an unknown command
gets sent over, an error would be raised. In spirit of the changes made
recently throughout to get rid of this, this part is changed to just
log a warning and send back an error if this happens.