flow_table: Fix bug on removing expired entries
The current implementation would just automatically make now equal to
the current time, even if some value was passed in. Normally this would
not matter because the entry is likely to be used without setting the
now value, but the unit tests caught this problem after attempting to
do so.