datapaths: Rewrote flow mod handling again
Code from the main _process_flow_mod method was shift to _rx_flow_mod
since there is no need to put it separately. The other methods had the
"_process" part of the name removed, so as to differentiate between
others that use this prefix and to shorten the header. The strict
parameter is now made an argument passed to modify and delete handling
to reduce checking.