datapaths: Implement emergency entry handling
Emergency entries must not have a timeout, so before anything else,
the entry must be checked and rejected if it does not adhere to this
condition. Also, the current switch does not have its own separate
emergency table, so such entries are not supported.

For future possible implementations of emergency entry handling, upon
deletion, an emergency entry should not generate a flow removal message
to the controller.