libopenflow: Add logging
It has long been an annoying issue that there was no logging in
libopenflow.  We didn't want to add it because in theory it should just
be for packing/unpacking.  We need have sort of optional logging -- an
external module can set a logger which will be used inside.

Additionally, we now set up such an additional logger in of_01.