libopenflow: Add some prereq awareness to matches
This is a sort of experimental patch.  It adds the fix() method to
of_match, which unsets fields which are unmatchable due to missing match
prerequisites.  See S3.4 in the OpenFlow 1.0.1 spec.

We currently do not call fix() be default, though we might some day.

This also a warning when prerequisites are not met, though this is only
shown when assertions are enabled.