packet.arp: Don't parse weird address types
When we see ARP addresses with strange address types (e.g., anything
besides Ethernet and IP), we no longer claim to have successfully
parsed the ARP packet.

Previously, we claimed to parse correctly, but didn't retain the
weird addresses, meaning that we could not then re-pack successfully.
We could easily have it store the weird addresses as raw bytes, so
that they could be re-packed, but most code probably assumes the
addresses are Ethernet and IP and it's not clear there are actual
use cases for anything else.