misc.arp_helper: More options for source MAC
The source MAC always used to be derived from the DPID, but using a custom
function to do so.  Firstly, this has been updated to use
Connection.eth_addr.  Secondly, there are now options to override this or
to use the *port* MAC instead.

This also adds some docstrings and breaks out some paramters to make the
code a bit more readable, and possibly fixes a bug in send_arp_reply().