misc.nat: Clean up and improve
I'd oversimplified some of the code I drew this from resulting in flows
getting stuck when the inside was quiet.  The old code was pretty bad
anyway, so this rewrites it all to hopefully be a bit cleaner.

Also fixes an issue where we were telling hosts inconsistent things about
our MAC address.