libopenflow: Improve flow_mod packing
First, this inserts a log statement where we'd wanted one before but
wasn't possible until we recently got logging in libopenflow finally.

Secondly, when using the .data attribute, we no longer transform it into
a .buffer_id, since this could lead to slightly surprising results if the
user wasn't expecting it.