Attempt to detect whether to launch of_01
We used to always launch of_01 (the OpenFlow 1.0 task) unless --no-openflow
was passed.  We now only do this if we notice that someone has checked for
or used core.openflow.  Note that openflow always loads in any case, it's
just a matter of whether of_01 does too.

In some cases, this may mean you need to put openflow or of_01 on the
commandline explicitly where you didn't used to.

There's actually a more radical patch which makes autoloading generic,
which may eventually get mainlined, but this commit is simpler and can be
seen as just a step toward ensuring that older components get updated to
do the right thing with respect to launching openflow/of_01.