recoco: Add unthreaded mode for SelectHub
Previously, SelectHub always ran in its own thread to put it on exactly
the same footing as any other IO loop.  However, this has a performance
penality, multiple IO loops have rarely been used in practice, and in any
case, SelectHub is probably special.

This patch adds the ability for SelectHub's functionality to be run
directly in the scheduler thread.  It is similar in spirit to changes to
SelectHub which are lumped into ed3b14b3c0 in the debugger branch, but
reuses more code, refactors part of the scheduler to be cleaner, and will
hopefully work correctly even with multiple threads.

Note that the old behavior (hopefully unchanged) is still the default.
This is pretty experimental.